20230203 migrations (#242)
* migrated PR#1183 - for new chains

* migrated PR#1168 - requester passing reserved parameter value

* migrated PR#1163 - new _gasPrice reserved parameter

* migrated PR#1161 - fetch-files command

* Broken links

* Update flex indexes
